refactor: update ID properties to long for larger values
Changed the `Id` and `AuthorId` properties in the `Author`, `Book`, and `BookCopy` classes from `int` to `long`. This update accommodates larger identifier values and ensures consistency across related entities in the model.

+++ LibraryAPI/Models/Book.cs
@@ -2,10 +2,10 @@ namespace LibraryAPI.Models;
 
 public class Book
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public string Title { get; set; } = string.Empty;
     public int Year { get; set; }
-    public int AuthorId { get; set; }
+    public long AuthorId { get; set; }
     public Author Author { get; set; } = null!;
     public ICollection<BookCopy> BookCopies { get; set; } = new List<BookCopy>();
 }
