+++ LibraryAPI/Models/BookCopy.cs
@@ -2,7 +2,7 @@ namespace LibraryAPI.Models;
 
 public class BookCopy
 {
-    public int Id { get; set; }
-    public int BookId { get; set; }
+    public long Id { get; set; }
+    public long BookId { get; set; }
     public Book Book { get; set; } = null!;
 }
